+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/data/dto/StatLegendDto.kt
@@ -57,5 +57,5 @@ data class StatLegendDto(
     @SerialName("level")
     val level: Int,
     @SerialName("xp_percentage")
-    val xpPercentage: Float,
+    val xpPercentage: Double,
 )
